Refactor demo metadata sourcing
