+++ playground/src/demos/types.ts
@@ -1,0 +1,11 @@
+import type { Component } from 'vue'
+import type { PropertyDefinition } from '@/library/types'
+
+export interface ComponentDemoConfig {
+  properties: PropertyDefinition[]
+  defaultProps?: Record<string, unknown>
+}
+
+export interface ComponentDemo extends ComponentDemoConfig {
+  component: Component
+}
